+++ Assets/Scripts/UI/Objects/PythonEditor.cs
@@ -1,6 +1,8 @@
 using System;
 using System.Collections;
+using System.Collections.Generic;
 using System.IO;
+using System.Text.RegularExpressions;
 using System.Threading;
 using System.Threading.Tasks;
 using TMPro;
@@ -19,11 +21,13 @@ public class PythonEditor : InitializeBehaviour
     public TextMeshProUGUI filename;
     public Scrollbar outputScrollBar;
 
+    Dictionary<string, GameObject> importLines;
     Thread execution;
     IPython engine;
 
     public override IEnumerator innerStart()
     {
+        importLines = new Dictionary<string, GameObject>();
         engine = GetComponent<IPython>();
         name = Path.GetFileName(fullpath);
         filename.text = name;
@@ -33,6 +37,53 @@ public class PythonEditor : InitializeBehaviour
     public override void innerBehaviour()
     {
         base.innerBehaviour();
+
+        // Cleaning previous imports
+        foreach (string import in importLines.Keys)
+        {
+            GameObject editor = GameObject.Find(import);
+            if (!editor)
+            {
+                Destroy(importLines[import].gameObject);
+                importLines.Remove(import);
+            }
+        }
+
+        // Check for imports
+        MatchCollection imports = Regex.Matches(pythonCode.text, @"import ([a-zA-Z0-9]+)");
+        foreach(Match import in imports)
+        {
+            string importedFile = import.Groups[1].Value + ".py";
+            GameObject editor = GameObject.Find(importedFile);
+
+            if (editor)
+            {
+                LineRenderer line;
+
+                if (!importLines.ContainsKey(importedFile))
+                {
+                    importLines.Add(importedFile, new GameObject("Line", typeof(LineRenderer)));
+                    line = importLines[importedFile].GetComponent<LineRenderer>();
+                    line.material = Instantiator.Instance.lineRendererMaterial;
+                    line.startWidth = .002f;
+                    line.endWidth = .002f;
+                }
+                else
+                {
+                    line = importLines[importedFile].GetComponent<LineRenderer>();
+                }
+                
+                line.SetPosition(0, transform.Find("Panel/Toolbar").position);
+                line.SetPosition(1, editor.transform.Find("Panel/Toolbar").position);
+            }
+        }
+    }
+
+    public override void onClose()
+    {
+        base.onClose();
+        foreach (GameObject import in importLines.Values)
+            Destroy(import);
     }
 
     public void Run()
@@ -57,7 +108,10 @@ public class PythonEditor : InitializeBehaviour
 
         await Task.Run(() =>
         {
-            try { execution.Join(); }
+            try 
+            { 
+                execution.Join(); 
+            }
             catch { }
         });
 
@@ -92,7 +146,6 @@ public class PythonEditor : InitializeBehaviour
             try 
             {
                 logText.text += engine.writer.GetContentFromBuffer();
-                logText.text += engine.errorWriter.GetContentFromBuffer();
                 outputScrollBar.value = 1f;
             }
             catch (InvalidOperationException)